Create Import Excel data into RedCap
I wanted to use the PYCap Python wrapper for Recap to read an excel file,  use mapping to create another excel file and also load the mapped values into RedCap 
I could not find an example anywhere  to create the ArrayList of Dictionaries that PyCap expects when importing more than one record.  Thought it would be good to give an example of how to load an excel file into a list of Dictionaries and the import the list into RedCap. I've got examples of handling datetime values, checkboxes, dropdown values etc  Will tidy up this example in time to come.